UBEER-569 Set spawnpointId on game creation not playfield load + destroy gamecontrollersignals when scene is destroyed + finish holder contruction before checking for account data in main menu

+++ ProjectUnityBeerServer/ProjectModules/ProjectOnlineGameModule/NetworkPackets/ClientNetworkPackets/ClientReceivedGameDataNetworkPacket.cpp
@@ -35,8 +35,7 @@ BinaryStream* ClientReceivedGameDataNetworkPacket::GetDataStream()
     datastream->WriteUInt8(onlinePlayer->GetLives());
     datastream->WriteUInt16(onlinePlayer->GetPositionX());
     datastream->WriteUInt16(onlinePlayer->GetPositionY());
-
-    datastream->WriteUInt16(onlinePlayer->GetSpawnPointId());
+    datastream->WriteUInt32(onlinePlayer->GetSpawnPointId());
   }
 
   return datastream;
